implement InterceptorPattern for NetworkRequest
